Update README, change namespace/package
