+++ src/Debugger/Debugger.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Redbox.Serilog.Stackdriver\Redbox.Serilog.Stackdriver.csproj" />
+    <ProjectReference Include="..\Raileasy.Serilog.Stackdriver\Raileasy.Serilog.Stackdriver.csproj" />
   </ItemGroup>
 
   <ItemGroup>
