+++ src/Debugger/appsettings.json
@@ -16,7 +16,7 @@
       {
         "Name": "Console",
         "Args": {
-          "formatter": "Redbox.Serilog.Stackdriver.StackdriverJsonFormatter, Redbox.Serilog.Stackdriver"
+          "formatter": "Raileasy.Serilog.Stackdriver.StackdriverJsonFormatter, Raileasy.Serilog.Stackdriver"
         }
       }
     ],
