--- src/Redbox.Serilog.Stackdriver/Redbox.Serilog.Stackdriver.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <PackageId>Redbox.Serilog.Stackdriver</PackageId>
-    <Authors>Redbox</Authors>
-    <Description>A Serilog Formatter for Stackdriver using JSON logs</Description>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <Copyright>Copyright 2019 (c) Redbox Automated Retail, LLC</Copyright>
-    <PackageTags>serilog stackdriver json log logging google</PackageTags>
-    <RepositoryUrl>https://github.com/redboxllc/stackdriver-serilog.git</RepositoryUrl> 
-    <RepositoryType>git</RepositoryType>
-    <PublishRepositoryUrl>true</PublishRepositoryUrl>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Serilog.AspNetCore" Version="8.0.0" />
-    <PackageReference Include="Serilog.Formatting.Compact" Version="2.0.0" />
-  </ItemGroup>
-
-
-
-</Project>
